fix test, separate recipe and ingredient test data
